requirements.txt
pyyaml
pytest
requests
beautifulsoup4